LayoutTests/fast/images/image-error-event-not-firing-expected.txt

CONSOLE MESSAGE: Refused to load http://www.myfakesiteabc.com/image.png because it does not appear in the img-src directive of the Content Security Policy.
This tests onerror event is fired if we change image src to a url blocked by content-security-policy.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error event fired.
PASS successfullyParsed is true

TEST COMPLETE
